flyttet interface til riktig sted
